requirements.txt
# Brain Capital - Copy Trading Platform
# Install: pip install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-socketio>=5.3.6
werkzeug>=3.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Binance API
python-binance>=1.0.19

# Security & Encryption
cryptography>=41.0.7

# Configuration
python-dotenv>=1.0.0

# Task Queue (ARQ - Async Redis Queue)
redis>=5.0.1
arq>=0.25.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0  # Async HTTP client for payment gateway

# Real-time Communication
eventlet>=0.34.2
gevent>=23.9.1

# Production Server (Windows)
waitress>=2.1.2

# Production Server (Linux/Docker)
gunicorn>=21.2.0

# Telegram Notifications & Bot
python-telegram-bot>=20.7

# Two-Factor Authentication (OTP)
pyotp>=2.9.0

# FastAPI & Exchange Support
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
ccxt>=4.1.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Observability (Prometheus + Loki)
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Web Push Notifications (PWA)
pywebpush>=1.14.0

# Twitter/X API for social sharing
tweepy>=4.14.0

# RAG Support Bot (LangChain + OpenAI + pgvector)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-postgres>=0.0.6
openai>=1.0.0
pgvector>=0.2.0
tiktoken>=0.5.0

# GeoIP for compliance (geo-blocking)
geoip2>=4.8.0

# Image Processing (Banner Generation for Influencer Dashboard)
Pillow>=10.2.0